# api/swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "API for SHOP Website",
    "description": "This is a REST API application made by ZDiDane for SHOP Website. \n\nIf you want to use this API, please get csrfToken first, then add csrfToken to header of request. \n\nIf you want to use API for Admin, please add bearAuth to header of request.",
    "contact": {
      "name": "ZDiDane",
      "email": "didannguyen17@gmail.com",
      "url": "https://didan.id.vn"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api"
    },
    {
      "url": "https://localhost:3000/api"
    },
    {
      "url": "http://shop.didan.id.vn/api"
    },
    {
      "url": "https://shop.didan.id.vn/api"
    }
  ],
...
  "paths": {
...
    "/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login",
        "description": "Endpoint to login.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "description": "Unprocessable Entity"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "csrfToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/SignIn"
              }
            }
          }
        }
      }
    },
...
    "/signup": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Signup",
        "description": "Endpoint to signup.",
        "responses": {
          "201": {
            "description": "Created"
          },
          "422": {
            "description": "Unprocessable Entity"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "csrfToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/SignUp"
              }
            }
          }
        }
      }
    },
    "/reset": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Get token to reset password",
        "description": "Endpoint to get token reset password.",
        "responses": {
          "201": {
            "description": "Created"
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "description": "Unprocessable Entity"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "csrfToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Reset"
              }
            }
          }
        }
      }
    },
...
    "/update-password": {
      "patch": {
        "tags": [
          "Auth"
        ],
        "summary": "Update password",
        "description": "Endpoint to update password.",
        "responses": {
          "201": {
            "description": "Created"
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "description": "Unprocessable Entity"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "csrfToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePassword"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "csrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token"
      },
      "bearAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "SignUp": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "anv"
          },
          "email": {
            "type": "string",
            "example": "anv@gmail.com"
          },
          "password": {
            "type": "string",
            "example": "123456"
          },
          "re_password": {
            "type": "string",
            "example": "123456"
          }
        },
        "required": [
          "name",
          "email",
          "password",
          "re_password"
        ],
        "xml": {
          "name": "SignUp"
        }
      },
      "SignIn": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "anb@gmail.com"
          },
          "password": {
            "type": "string",
            "example": "123456"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "xml": {
          "name": "SignIn"
        }
      },
      "Reset": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "anv@gmai.com"
          }
        },
        "required": [
          "email"
        ],
        "xml": {
          "name": "Reset"
        }
      },
...
    }
  }
}